pr: refactored config vars
